Change file attributes to uint8_t

+++ src/kernel/filesystem/utils/api_utils.h
@@ -5,5 +5,5 @@
 
 namespace api_utils {
 
-	bool Check_File_Attributes(std::uint16_t file_attributes, kiv_os::NFile_Attributes attribute);
+	bool Check_File_Attributes(std::uint8_t file_attributes, kiv_os::NFile_Attributes attribute);
 }
