+++ src/kernel/filesystem/vfs_layer/vfs_element.cpp
@@ -18,7 +18,7 @@ void VFS_Element::Close() {
 	throw std::runtime_error("Unsupported operation CLOSE.");
 }
 
-bool VFS_Element::Set_File_Attributes(std::uint16_t file_attributes) {
+bool VFS_Element::Set_File_Attributes(std::uint8_t file_attributes) {
 	throw std::runtime_error("Unsupported operation SET FILE ATTRIBUTES.");
 }
