+++ src/kernel/filesystem/vfs_layer/vfs_fat_element.cpp
@@ -5,11 +5,11 @@ void VFS_Fat_Element::Close() {
 	//EMPTY
 }
 
-bool VFS_Fat_Element::Is_Convertable(std::uint16_t file_attributes) {
+bool VFS_Fat_Element::Is_Convertable(std::uint8_t file_attributes) {
 	throw std::runtime_error("Unsupported operation IS_CONVERTABLE.");
 }
 
-bool VFS_Fat_Element::Set_File_Attributes(std::uint16_t file_attributes) {
+bool VFS_Fat_Element::Set_File_Attributes(std::uint8_t file_attributes) {
 	if (this->Is_Convertable(file_attributes)) {
 		this->file_attributes = file_attributes;
 		return true;
