+++ src/kernel/filesystem/vfs_layer/vfs_fat_element.h
@@ -5,12 +5,12 @@
 
 class VFS_Fat_Element : public VFS_Element {
 protected:
-	std::uint16_t file_attributes;
+	std::uint8_t file_attributes;
 	char file_name[12]; //TODO change to std::array<char, MAX_FILENAME_SIZE> (na hodne mistech)
 	std::uint64_t file_position;
 
-	virtual bool Is_Convertable(std::uint16_t file_attributes);
+	virtual bool Is_Convertable(std::uint8_t file_attributes);
 public:
 	virtual void Close();
-	virtual bool Set_File_Attributes(std::uint16_t file_attributes);
+	virtual bool Set_File_Attributes(std::uint8_t file_attributes);
 };
